requirements.txt
streamlit
pandas
numpy
tensorflow-cpu
scikit-learn
joblib
scikit-learn==1.6.1